+++ src/word_count.h
@@ -16,7 +16,7 @@
 #include <boost/locale.hpp>
 #include <boost/version.hpp>
 
-std::unordered_map<std::string, u_int> stringToMap(std::string const& text, std::unordered_map<std::string, u_int> *res);
-void printMap(std::unordered_map<std::string, u_int> const& word_map);
-void sortMap(std::unordered_map<std::string, u_int> const& word_map, int first_n=-1);
+std::unordered_map<std::string, size_t> stringToMap(std::string const& text, std::unordered_map<std::string, size_t> *res);
+void printMap(std::unordered_map<std::string, size_t> const& word_map);
+void sortMap(std::unordered_map<std::string, size_t> const& word_map, int first_n=-1);
 #endif //UNTITLED4_WORD_COUNT_H
